Move mux functionality into CondSwap chip
